fix queries:
A record is set up so that the feature (subject) is linked to an outcome (object). Therefore the correct query setup should have the input ID as the subject and the outcome as the object.

 input ID should be the subject, data and output ID should be object

also, UNII IDs don't need the extra escape character. the queries as-is fail since then extra \\\\ are made when the query is constructed